+++ app/Http/Controllers/PostsApiController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Post;
 use App\Models\Comment;
+use App\Models\Category;
 use Illuminate\Support\Facades\Auth;
 
 class PostsApiController extends Controller
@@ -15,18 +16,23 @@ class PostsApiController extends Controller
     }
 
     public function getPost($postid){
+
         if(Post::where('id', $postid)->exists()){
+
             $post = Post::where('id', $postid)->get();
             return response($post, 200);
+
         } else {
-            return response()->json([
-                "message" => "Post not found"
-            ], 404);
+
+            return response()->json(["message" => "Post not found"], 404);
+
         }
     }
     
     public function getPostComments($postid, Comment $comments){
+
         if(Post::where('id', $postid)->exists()){
+
             return response(array(
                 $post = Post::where('id', $postid)->get(),
                 $comments = Comment::where('post_id', $postid)->get()), 200);
@@ -49,11 +55,17 @@ class PostsApiController extends Controller
         if(! $isGuest){
             $user_id = auth()->user()->id;
 
-            return Post::create([
-                'title' => request('title'),
-                'content' => request('content'),
-                'category_id' => request('category_id'),
-            ]);
+            if (Category::where('id', request('category_id'))){
+
+                return Post::create([
+                    'title' => request('title'),
+                    'content' => request('content'),
+                    'category_id' => request('category_id'),
+                    'user_id' => $user_id,]);
+            }
+            else{
+                return response()->json(["message" => "Category not found"], 404);
+            }
         }
         else{
             return response()->json([
@@ -78,25 +90,22 @@ class PostsApiController extends Controller
                 if($user_id == $post -> user_id || $user_role == 1){
                     $post->title = is_null($request->title) ? $post->title : $request->title;
                     $post->content = is_null($request->content) ? $post->content : $request->content;
-                    $post->category_id = is_null($request->category_id) ? $post->category_id : $request->category_id;
+                    $post->category_id = $post->category_id;
                     $post->user_id = $post->user_id;
                     $post->save();
                     
-                    return response()->json([
-                        "message" => "Post update finished",
-                        "post" => $post
-                    ], 200);
-                }else{
-                    return response()->json([
-                        "message" => "Post not found"
-                    ], 404);
+                    return response()->json(["message" => "Post update finished", "post" => $post], 200);
+                }
+                else{
+                    return response()->json(["message" => "Unauthorized"], 401);
                 }
             }
             else{
-                return response()->json([
-                    "message" => "Post missing"
-                ], 401);
+                return response()->json(["message" => "Post missing"], 404);
             }
+        }
+        else{
+            return response()->json(["message" => "Unauthorized"], 401);
         }
     }
 
@@ -110,22 +119,31 @@ class PostsApiController extends Controller
 
             if(Post::where('id', $postid)->exists()){
                 $post = Post::find($postid);
-                $post->delete();
-    
-                return response()->json([
-                    "message" => "Post removed"
-                ], 202);
-            } 
-            else{
-                return response()->json([
-                    "message" => "Post missing"
-                ], 404);
+                if ($user_id == $post->user_id || $user_role == 1)
+                {
+
+                    $post->delete();
+
+                    return response()
+                        ->json(["message" => "post deleted"], 202);
+                }
+                else
+                {
+                    return response()
+                        ->json(["message" => "Unauthorized"], 401);
+                }
+
+            }
+            else
+            {
+                return response()
+                    ->json(["message" => "post not found"], 404);
             }
         }
-        else{
-            return response()->json([
-                "message" => "Unauthorized"
-            ], 401);
+        else
+        {
+            return response()
+                ->json(["message" => "Unauthorized"], 401);
         }
     }
 }
